Remove process_document_file
